Workload and data generator for Genome Benchmark

+++ include/data/generators/genome_generator.hpp
@@ -9,18 +9,18 @@
 class GenomeGenerator : public AbstractGenerator{
     public:
         GenomeGenerator(
-                int64_t n_of_rows_,
-                int64_t number_of_queries_,
-                int64_t query_type_,
+                size_t n_of_rows_,
+                size_t number_of_queries_,
+                size_t query_type_,
                 const std::string &feature_vectors_file_,
                 const std::string &genes_file_
                 );
         void generate(std::string table_path, std::string workload_path) override;
 
     private:
-        int64_t n_of_rows;
-        int64_t number_of_queries;
-        int64_t query_type;
+        size_t n_of_rows;
+        size_t number_of_queries;
+        size_t query_type;
         std::string FEATUREVECTORS_FILE= "chr22_feature.vectors";
         std::string GENES_FILE = "genes.txt";
 };
